Move docker commands in their own file
